requirements.txt
openai==1.30.1
psycopg2-binary
gradio==3.50.2
python-dotenv